litedb

+++ samples/aspnetapp/aspnetapp/Models/QuestionNode.cs
@@ -1,11 +1,12 @@
 ﻿using System.Collections.Generic;
+using LiteDB;
 
 namespace aspnetapp.Models
 {
     public class QuestionNode
     {
         public int Id { get; set; }
-        public int LastUsed { get; set; }
+        public long LastUsed { get; set; }
         public byte[] Image { get; set; }
         public string Header { get; set; }
         public string Question { get; set; }
